# docs/contracts/openapi-v1.yaml
openapi: 3.1.0
info:
  title: PoolAI Release 1 API
  version: 1.0.0
  summary: PoolAI 非盈利版 Release 1 的可执行 HTTP 契约
  description: |
    本文件冻结 Release 1 的公开数据面和最小控制面。数据面只承诺
    OpenAI/Codex 的文本与函数工具垂直切片，不宣称覆盖完整 OpenAI API。

    管理输入与单请求 OpenAI usage 使用 JSON number，范围为 `0..9007199254740991`。
    Group quota、时间窗口和 Account/Group 聚合可能因长期累加或超额核销超过该范围，
    因此一律使用规范十进制 string；服务端使用 .NET `BigInteger` / PostgreSQL `NUMERIC`。

    `/v1/responses` 的流遵循 OpenAI typed SSE：`event` 与 `data.type` 相同，
    `sequence_number` 从 0 开始严格递增，正常流以 `response.completed` 结束且不发送
    Chat Completions 风格的 `[DONE]`。流已开始后的错误发送 `event: error` 后关闭连接，
    不再发送 `response.completed`。参考：
    https://developers.openai.com/api/docs/guides/streaming-responses 与
    https://developers.openai.com/api/reference/resources/responses/streaming-events 。

    控制面资源变更必须携带 `Idempotency-Key`。认证会话端点由一次性 Refresh Token
    轮换保证重放安全，因此不要求该 Header。所有 versioned resource 的状态变更、
    PATCH、软退役和 Group 配额调整/重置还必须携带 `If-Match`。`/v1/*` 错误只使用
    `application/json` GatewayProblem；`/api/v1/*` 错误只使用
    `application/problem+json` Problem Details。两个 Schema 在机器契约中彻底隔离：
    控制面 ControlPlaneProblem 不允许 OpenAI `error` 投影，数据面 GatewayProblem
    必须携带该投影。统一错误目录与 phase-aware retry/failover 矩阵见
    `error-catalog.md`。
  license:
    name: Internal specification
servers:
  - url: https://poolai.example.com
    description: 生产环境示例；部署时替换
...
paths:
...
  /api/v1/admin/accounts:
    get:
      tags: [AdminAccounts]
      operationId: adminListAccounts
      summary: 分页列出上游 Account
      x-required-roles: [admin, operator, auditor]
      security:
        - UserJwt: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 不返回凭据明文
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [AdminAccounts]
      operationId: adminCreateAccount
      summary: 创建 OpenAI API-key Account
      x-required-roles: [admin, operator]
      security:
        - UserJwt: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
      responses:
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '201':
          description: Account 以 status=disabled、health.status=unknown、active_leases=0 创建；credential 只写不回显
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/admin/accounts/{accountId}:
    get:
      tags: [AdminAccounts]
      operationId: adminGetAccount
      summary: 查询 Account
      x-required-roles: [admin, operator, auditor]
      security:
        - UserJwt: []
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account 详情，不返回凭据明文
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    patch:
      tags: [AdminAccounts]
      operationId: adminUpdateAccount
      summary: 更新 Account；可轮换 credential
      x-required-roles: [admin, operator]
      security:
        - UserJwt: []
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '200':
          description: 更新后的 Account
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [AdminAccounts]
      operationId: adminRetireAccount
      summary: 软退役 Account
      description: |
        仅将状态转为 terminal `retired`，不物理删除 credential 密文元数据、binding 或 usage。
        若 Account 仍存在 enabled Supply binding 则返回 `account_in_use`；已禁用/历史引用不阻止退役。
      x-required-roles: [admin, operator]
      security:
        - UserJwt: []
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/XChangeReason'
      responses:
        '204':
          description: 已软退役；历史依赖保留
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            ETag:
              $ref: '#/components/headers/ETag'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/admin/usage/accounts:
    get:
      tags: [AdminAccounts]
      operationId: adminListAccountUsage
      summary: 按 Account、Group、模型与时间聚合 Token
      x-required-roles: [admin, operator, auditor]
      security:
        - UserJwt: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: account_id
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: group_id
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: model
          in: query
          schema: { type: string, maxLength: 200 }
        - name: from
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Timestamp' }
        - name: to
          in: query
          required: true
          schema: { $ref: '#/components/schemas/Timestamp' }
        - name: granularity
          in: query
          schema: { type: string, enum: [hour, day], default: hour }
      responses:
        '200':
          description: 跨账号聚合分页
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageAggregatePage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    PoolApiKey:
      type: http
      scheme: bearer
      bearerFormat: sk-…
      description: |
        数据面专用 API Key，放在 `Authorization: Bearer sk-...`。服务端只保存
        HMAC-SHA256 hash 与可展示 prefix。不能放在 URL、cookie 或 query string。
    UserJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        控制面短期 Access JWT，放在 `Authorization: Bearer <jwt>`，必须各包含唯一一个
        `sub`、`role`、`token_version` 和 `sid=<refresh family_id>`。每个 `UserJwt` 请求都必须在 Policy 前以同一 PostgreSQL
        强读确认该 User active、未删除、`token_version` 匹配、唯一 canonical role，以及
        该 family 仍有 active、未过期 refresh generation；JWT 中的 role 与 token_version
        只作为格式和并发版本候选，不能直接授权。任一条件不匹配统一返回 `invalid_user_token`；强读失败
        返回 `dependency_unavailable`，不能回退到 Redis 或 JWT。不能用 `sk-` Key 替代。
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: 服务端返回的不透明游标
      schema:
        type: string
        minLength: 1
        maxLength: 512
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        业务资源变更幂等键。ASCII 1..128 字符；作用域为 principal + method + path，
        至少保留 24 小时。相同键不同请求摘要返回 `idempotency_conflict`。
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
...
  headers:
    XRequestId:
      description: 服务端为本次入站请求生成的 UUIDv7 标识；日志、trace、错误 body 与 Header 保持一致。幂等重放也使用本次请求的新标识，不复用首次请求的关联 ID。
      schema:
        $ref: '#/components/schemas/Uuid'
    Location:
      description: 新资源的绝对或同源相对 URI
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: 缺少、无效或过期的凭据
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
    Forbidden:
      description: 凭据有效，但角色、订阅、Group 或模型授权不足
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
...
    PayloadTooLarge:
      description: 请求体超过入口或应用上限
      x-http-status: 413
      x-error-code: payload_too_large
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
    UnsupportedMediaType:
      description: 请求 Content-Type 不受支持
      x-http-status: 415
      x-error-code: unsupported_media_type
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
...
    TooManyRequests:
      description: 通用速率限制或 Control 进程级 bulkhead 拒绝
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
          examples:
            gatewayOverloaded:
              summary: "Control 令牌/有界队列已满；必须同时返回 Retry-After: 1"
              value:
                type: https://poolai.example/problems/gateway-overloaded
                title: Gateway overloaded
                status: 429
                detail: 当前控制面有界并发容量已满，请稍后重试。
                instance: /api/v1/admin/runtime-settings
                code: gateway_overloaded
                request_id: 0190f8bf-a040-7444-a2ca-c4bc32e48b47
                retryable: true
                retry_after_seconds: 1
...
    InternalError:
      description: 内部错误或 Token 精确整数溢出；不暴露堆栈、SQL、原始溢出值或秘密
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
...
    ServiceUnavailable:
      description: 配额真相库、硬协调或依赖服务暂不可用
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ControlPlaneProblem'
          examples:
            dependencyUnavailable:
              summary: "PostgreSQL 或其他必需依赖暂不可用；响应 Header 必须 Retry-After: 1"
              value:
                type: https://poolai.example/problems/dependency-unavailable
                title: Dependency unavailable
                status: 503
                detail: 必需依赖暂不可用，请稍后重试。
                instance: /api/v1/admin/groups
                code: dependency_unavailable
                request_id: 0190f8bf-a040-7444-a2ca-c4bc32e48b47
                retryable: true
                retry_after_seconds: 1
  schemas:
...
    AccountPage:
      type: object
      additionalProperties: false
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        page:
          $ref: '#/components/schemas/PageInfo'
...